server/requirements.txt
Flask>=2.3.0
Flask-Cors>=3.0.10
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
google-generativeai>=0.3.2
